Refactored code for setup of editing homepage features and descriptions
