+++ client/groups/config/groups.client.init-module.js
@@ -11,7 +11,7 @@ highChartsDrilldown(highCharts);
 highChartsTreeMap(highCharts);
 highChartsMore(highCharts);
 
-window.Highcharts = highCharts; 
+window.Highcharts = highCharts;
 
 (function () {
 	'use strict';
