+++ client/groups/directives/manage-homepage.client.directive.js
@@ -1,0 +1,19 @@
+(() => {
+	'use strict';
+
+	angular
+		.module('groups')
+		.directive('manageHomepage', manageHomepage);
+
+
+	function manageHomepage () {
+
+		const directive = {
+			restrict: 'E',
+			templateUrl: '/groups/views/manage-homepage.client.view.html'
+		}
+
+		return directive;
+	}
+
+})();
