+++ client/layout/views/layout-header.client.view.html
@@ -83,7 +83,7 @@
 
                   <li ng-if="user.isLoggedIn && user.currentUser.isAdmin"> <a ui-sref="dashboard"><i class="fa fa-bar-chart" aria-hidden="true"></i> Community Dashboard</a></li>
                   <li ng-if="user.isLoggedIn && user.currentUser.isAdmin"> <a ui-sref="manageGroupClassification"><i class="fa fa-database" aria-hidden="true"></i> Manage Group Classifications</a></li>
-                  <li ng-if="user.isLoggedIn && user.currentUser.isAdmin"> <a ui-sref="manageHomeSlider"><i class="fa fa-file-image-o" aria-hidden="true"></i> Manage Home Slider </a></li>
+                  <li ng-if="user.isLoggedIn && user.currentUser.isAdmin"> <a ui-sref="manageHomepage"><i class="fa fa-home" aria-hidden="true"></i> Manage Homepage </a></li>
 
                   <li ng-if="user.isLoggedIn"> <a ui-sref="createGroup"><i class="fa fa-plus" aria-hidden="true"></i> Create Group</a></li>
                   <!-- <li ng-if="user.isLoggedIn"> <a href="#"><i class="fa fa-cog" aria-hidden="true"></i> Settings</a></li> -->
